Fixed a bug in the help window where the help for the OpenGL demo was randomly displayed for all demos but the OpenGL demo :-). Details of the fix below.
(Switch_Page): <<if Get_Current_Page (Notebook) = 0>> has been changed
into <<if Get_Current_Page (Notebook) = 3>> since the page number of the
opengl demo is 3, not 0.

(Initialize): Connect the Switch_Page signal *after* (ie the After flag
of the "Connect" procedure is set to True). Otherwise, the callback is
invoked before gtk+ switches to the new page. The consequence was that
Get_Current_Page returned the page number of the previous page.