Fix GtkAda shared libraries installation on Windows.
It was not possible to build a shared libraries depending on GtkAda.
The GtkAda shared libraries was looked into the gtkada/relocatable
directory but was only installed into the <root>/bin directory by
the GtkAda InstallShield.

Do not copy the shared library into the gtkada/relocatable directory in
the Makefile to be in sync with what is done by the InstallShield. It
is better to keep developers setup close to what is done for end-users.
--
Work for I223-013.