* binding.xml: Add binding for GtkTargetFlags in Gtk.Enums. (GtkIconView): Adapt to new location of Target_Entry_Array.
Add bindings for GtkSelectionData, GtkTargetList, GtkTargetEntry.

* data.py:
Add bindings for GtkSelectionData, GtkTargetEntry, GtkTargetList.

* gdk-dnd.adb:
Adapt to changes in Gtk.Selection.

* gdk-dnd.ads:
Adapt to changes in Gtk.Selection.

* gtk-clipboard.adb:
Adapt to changes to Gtk.Selection.

* gtk-clipboard.ads:
Adapt to changes to Gtk.Selection.

* gtk-dnd.adb:
Adapt to changes to Gtk.Selection.

* gtk-dnd.ads:
Adapt to changes to Gtk.Selection.

* gtk-tree_dnd.adb:
Adapt to changes to Gtk.Selection.

* gtk-tree_dnd.ads:
Adapt to changes to Gtk.Selection.

* gtk-tree_view.adb:
Adapt to changes to Gtk.Selection.

* gtk-tree_view.ads:
Adapt to changes to Gtk.Selection.

* create_clipboard.adb:
Adapt to changes to Gtk.Selection.

* create_dnd.adb:
Adapt to changes to Gtk.Selection.

For LA22-029.