Move to Cocoa framework on OSX, enable full-screen mode for application windows.
* configure, Makefile.in, aclocal.m4, shared.gpr:
Enable objective-c source code build on OSX.

* misc.c, misc_osx.h, misc_osx.m:
Move most of the OSX-specific code to a separate Objective-C source file,
move to a full Cocoa implementation for the Open File event handler, and
add a utility function to enable full-screen mode on a specific GtkWindow.

* generate_prj.sh:
Move the framework dependency to Cocoa

* gtkada-application.ad[bs]:
Refine the API to add means to select specific added options provided
by the object.

For both MB07-006 and MB07-007.