Major review of the handling of Ada_GObject_Class
This checkin cleans up the use of that type, since there were some
confusion with GObject_Class.
Add a Class_Init parameter to Initialize_Class_Record, to separate the
creation of the gtk+ GObjectClass* and the GType.
This check in introduces minor backward incompatibilities because
Set_Default_*_Handler now take a GObject_Class as argument, not an
Ada_GObject_Class. This helps clarify they should be called from
Class_Init and not when Initialize_Class_Record is called, which is
also more efficient.
With this check in, it is now possible to create a new Ada widget
that implements the GtkScrollable interface.

N326-016