Redo the computation of link layout
This improves the support for orthogonal links, since it is now possible
to attach links to specific sides on the end objects, and thus this makes
the routing slightly more complex.
Add support for "orthocurve" link, which are bezier splines using the same
computation as orthogonal links for the control points.

N226-019