Change the way the pointer to AdaGObject class is stored
We no longer use pointer arithmetics (and g_type_query), but a more
efficient way, but using the padding array that gtk+ already has in
GObjectClass (and which it does not use).
This is more efficient.
The previous implementation for some reason was leading to corrupted
data for the canvas_view when a second signal was added to it, because
it also has customer properties getter.