canvas_view new event "item_event"
This is the first step towards user interaction in the canvas.
The new test shows how to handle clicking on item to perform action.
In particular, it is now also possible to change the text displayed
by a Text_Item, which highlighted a bug in the way the model's layout
was reused after we allow ellipsis, which impacts size computation.