Add support for overriding virtual-method for classes
This was only possible for interfaces, but in fact for a GApplication
we might need to override the local_command_line virtual_method, as
per the documentation.
Group all with statements for which we need a pragma Warnings(Off) so
that a single instance of the pragma is needed.

Minor style fixes

N625-006

Change-Id: I36aa3285fbabc964e198f1b1e2f0cf21a2210037