GtkAda.MDI.Child_Flags Float_As_Dialog -> Float_To_Main
Float_As_Dialog was renamed Float_To_Main. The difference in behavior
is that Float_As_Dialog attached the dialog to no window. Now the dialog
is attached to the main window.

This fixes attaching dialogs to transient MDIs that have shorter
lifetimes than the dialog, but also keeps the behavior of dialogs who
stays on top of the main application all the time, which is desired.

N410-050 - N523-026

Change-Id: I9231f8a7f076ac8cab72d4d4889d26e106495b10