Fix binding for Glib.Application's local_command_line handling
The type of parameter was wrong, thus it was not possible to get the
command line switches.

Provide a new procedure Glib.Option.Parse, similar to the one we
already had, that can be used in the local_command_line handling via
a Goption_Context. The version we already have is suitable for use in
the command_line handling (i.e. after the application has connected to
the GUI server).

Add support for overriding the type of parameters for virtual-function
in the python scripts.

Part of NA22-010

Change-Id: I34eb031150f0d2ebd73a44e3060b021593248c8a