Gtk.Dialog.No_Separator has been removed in some earlier version of gtk
It was in fact replaced with Use_Header_Bar, and this results in GtkAda
application potentially having a mix of dialog showing action buttons
in the header bar, rather than use the default settings value.

Add various subprograms to make it easier for applications to set up
dialog that use header bars.

NB28-004

Change-Id: Iedd4532596168aba6c02125151514ae23389dd81