(Save_Desktop): prevent error when MDI has no toplevel window
Although this is not used in practice, this clean up came up as part of
reviewing the exit sequence in GPS. Since we were trying to save the desktop
in the "destroy" callback for the main window, the mdi had been unparented
when Save_Desktop is called.
This does not quite work though when restoring, since we do not have a size
for the main window, and the MDI's saved size would result in setting a
minimum size, not a default size.

Part of NC11-021

Change-Id: Ic804a3ddbb1bc7ecf99e34d3fb2bc8d47194812e