Update comments
As part of a fix for O209-022 and O220-018.
This patch documents an approach that did not work fine, but which we might
be tempted to use again in the future.

This fixes the scenario in those tickets: give focus to editor, then to
 the Codepeer messages window, and press backspace. The character should
 not be deleted in the editor.
It also seems to keep the scenarios that were previously identified in this
 part of the code (805e05837f4aa683da616eb890329c60b3babe9a):
 - have a non-empty floating locations window. Give focus to a
   non-floating editor and press the key shortcut for "next tag". The cursor
   position should change, and the focus remain in the editor.
 - Have a floating Search window. Click in the main menu bar to make it
   non floating. The menu should apply to the search window. Same if the
   floating window is an editor and you select a menu like /Edit/Cut for
   instance.
It breaks contextual menus in GPS, though

Change-Id: I71825db08db122c179709e11d307a49156c8c481