Add a mechanism to avoid relying on chance
When a child is destroyed interactively, Close_Child
gives the focus to the previous child, then unfloats
the child being closed: this operation assigns the
focus to the child being unfloated, and then relies
on gtk behavior to give the focus back to the previous
child. This may not always happen: add a mechanism
to control the child we focus.

Change-Id: Id2603a67811206c297e90dce300bc5767956385a