Add support for width and height proportional to parent
It is now possible to indicate that a child occupies 50% of the
parent container for instance. We used to make all children (but
floating ones) 100% of the parent, but this limits the layouts
that can be done. For instance, drawing a rectangle with a specific
width, and below it a label with a larger width was hard to
achieve except when hard-coding coordinates.
Other types of units were introduced for Set_Size, to control whether an
item has its preferred, size, the size of its parent, or fits in its
parent including its margins.

Clarify roles of Anchor_[XY] and Alignment, and when they apply.
When an item has an anchorx of 1.0, it means its right edge is aligned with
the right edge of its parent. So we need to compute the position of the left
edge before we can recompute the width of the item.
This was removing the possibility of setting an outPort on a block which
would be aligned on the right edge of the block

(Set_Size_Range): split into Set_Width_Range and Set_Height_Range, since
calling these will override explicit sizes set by the user, and he might
not want to reset both axix.

(Initialize_Details): new procedure


Part of O917-061 (1/3)

Change-Id: I9951fe148baad60c7498e99899cedb45312de3d3