(Set_Menu_Model): new subprogram
The /Window menu can now be described as a menu model rather than with
widgets. This makes it possible to duplicate this menu in floating windows,
or even to show it in system menus on Ubuntu Unity and OSX.

("unfloat_child"): new signal on the MDI

(Create_Menu): it is now possible to create multiple instances of the
menu. The MDI no longer stores a reference to them, instead the menu
items connect to the proper signals to change their status. This is useful
to have multiple menubars (for instance on floating windows), when not
using Set_Menu_Model.


G222-043 (1/2) and N829-014

Change-Id: If8bf784954eb649cccb854724f13e26d5beb67ad