Fix memory leaks
In a number of cases, gtk+ accepts user data, but does not provide a
callback to indicate when this data can be freed. GtkAda was still
allocating memory, though, so this resulted in memory leaks. In some
cases, we were able to fix the binding to immediately free the data,
no longer needed by gtk+ (typically Foreach procedures or similar
iterators).
However, in some cases the lifetime of the data is unknown. Rather
than have known memory leaks, we chose to stop binding the
corresponding functions (note that we still provide a binding
to the function, with no user data -- only the version with user
data was removed). Here is the full list:

     gtk.accel_group.find_user_data
     gtk.builder.connect_signals_full_user_data
     gtk.clipboard.request_contents_user_data
     gtk.clipboard.request_image_user_data
     gtk.clipboard.request_rich_test_user_data
     gtk.clipboard.request_targets_user_data
     gtk.clipboard.request_text_user_data
     gtk.clipboard.request_uris_user_data
     gtk.main.key_snooper_install_user_data
     pango.attributes.filter_user_data

OC10-015

Change-Id: I9d62065d263cf4f5ea9ba2bbddbadf0648eff7a4