Use a CSS class to highlight notebook MDI children
P303-019

The Gtkada.MDI.Highlight_Child procedure now uses CSS classes
to highlight and unhighlight a notebook MDI child: The
"mdi-highlighted-tab" class is added to the child's tab label
when highlighting it and removed when unhighlighting it.

This allows to customize the child tab's label appearance when
highlighted using CSS directly.

Change-Id: Ia56699438018e2a799610b665533de6b00c5d35e