Add an internal mechanism to freeze and thaw focus
This avoids too many focus changes when restoring a desktop
or loading perspectives.

Fix All-floating mode: since focus is given asynchronously
by Gtk, and since the MDI sets the menu in reaction to who
has the focus, and since toggling the menu gives the focus
to a child, there could be an infinite loop when activating
this mode.

Follow up on P323-023.
Change-Id: I86f8026833d3c8119327d7b7497bfcfc0235a1a6