("inline_editing_started", "inline_editing_finished"): new signals
New signals added to the canvas view, so that GPS can invalidate the
current context whenever something occurs on the canvas that might
change which actions are applicable.

When an item is edited inline, and moved at the time, we now also
move the editing widget to keep overlapping the item.

Change-Id: I458ef32a90e607aa3e86bd588d1c0ea62cbfa69e