(Check_Interactive_Selection_Dialog): handle multiple panes
In emacs, Cx-o should move to the next pane, including when we have
three or more opened. The previous implementation here would only
select between the current and the next pane (and back) because of the
way we depend on the MDI.Items list.
Instead, we need to traverse directly at the notebook level.

Fixes Q130-008


Change-Id: I119c71468700bba320eaf1d8275031071524e423